separate proxy and test structure correctly
