+++ Dockerfile-proxy
@@ -8,20 +8,22 @@ RUN apk add --no-cache --update git build-base ca-certificates zstd
 
 RUN mkdir config
 COPY config/ ./config
-COPY . .
+COPY spawn_proxy/ ./spawn_proxy
+COPY apps/ ./apps
 COPY mix.exs .
 COPY mix.lock .
 
 RUN mix local.rebar --force \
     && mix local.hex --force \
     && mix deps.get \
-    && mix deps.get \
     && mix release.init
 
-RUN echo "-name spawn@${HOSTNAME}" >> ./rel/vm.args.eex \
+RUN echo "-name proxy@${HOSTNAME}" >> ./rel/vm.args.eex \
     && echo "-setcookie ${NODE_COOKIE}" >> ./rel/vm.args.eex
 
-RUN MIX_ENV=prod mix release proxy
+RUN cd spawn_proxy/proxy \
+    && mix deps.get \
+    && mix release proxy
 
 # ---- Application Stage ----
 FROM alpine:3
