+++ apps/actors/test/actors/actors_test.exs
@@ -163,7 +163,7 @@ defmodule ActorsTest do
     end
 
     # Useful for distributed testing due to natural eventuality problems
-    defp loop_until_ok(func, timeout \\ 15_000) do
+    defp loop_until_ok(func, timeout \\ 10_000) do
       task = Task.async(fn -> do_loop(func, func.()) end)
 
       Task.await(task, timeout)
@@ -173,9 +173,9 @@ defmodule ActorsTest do
 
     defp do_loop(func, _resp) do
       do_loop(func, func.())
-    catch
-      _ -> do_loop(func, func.())
     rescue
+      _ -> do_loop(func, func.())
+    catch
       _ -> do_loop(func, func.())
     end
   end
