--- apps/proxy/test/test_helper.exs
@@ -1,3 +1,0 @@
-ExUnit.start()
-
-Spawn.InitializerHelper.setup()
