+++ apps/sidecar/mix.exs
@@ -14,7 +14,8 @@ defmodule Sidecar.MixProject do
       lockfile: "../../mix.lock",
       elixir: "~> 1.14",
       start_permanent: Mix.env() == :prod,
-      deps: deps()
+      deps: deps(),
+      elixirc_paths: elixirc_paths(Mix.env())
     ]
   end
 
@@ -34,4 +35,7 @@ defmodule Sidecar.MixProject do
       {:statestores, "~> 0.1", in_umbrella: true}
     ]
   end
+
+  defp elixirc_paths(:test), do: ["lib", "test/support"]
+  defp elixirc_paths(_), do: ["lib"]
 end
