+++ config/config.exs
@@ -50,8 +50,4 @@ config :protobuf, extensions: :enabled
 #  registry: :default,
 #  auth: false
 
-# App Configuration
-config :proxy,
-  http_port: System.get_env("PROXY_HTTP_PORT", "9001") |> String.to_integer()
-
 import_config "#{config_env()}.exs"
