+++ spawn_proxy/proxy/config/config.exs
@@ -1,0 +1,7 @@
+import Config
+
+config :proxy,
+  http_port: System.get_env("PROXY_HTTP_PORT", "9001") |> String.to_integer()
+
+import_config "../../../config/config.exs"
+import_config "#{config_env()}.exs"
