+++ spawn_proxy/proxy/lib/proxy/application.ex
