+++ spawn_proxy/proxy/lib/proxy/routes/health.ex
